LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/background-position-expected.txt

CONSOLE MESSAGE: Error: '<position>' is not a valid CSS component

Harness Error (FAIL), message = Error: '<position>' is not a valid CSS component

FAIL Can set 'background-position' to CSS-wide keywords assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to var() references assert_equals: expected 2 but got 1
